sigmapiweb/requirements/github.txt
-r base.txt
pylint
pylint-django
pep8
